+++ src/Enums.ts
@@ -1,10 +1,87 @@
-// TODO: copy docs page from menu-admin-client
+/**
+See [Enums considered harmful](https://www.youtube.com/watch?v=jjMbPt_H3RQ) for the motivation
+behind this custom type. (Also worth noting is that in TypeScript 5.0 [all Enums are considered
+unions](https://devblogs.microsoft.com/typescript/announcing-typescript-5-0/#all-enums-are-union-enums).)
+
+Similar to [variants](variants.md) for disciminated unions, `enumOf` allows you to more easily
+and safely create and work with enums in TypeScript.
+
+## Basic Example
+
+```ts
+export const MyEnum = enumOf(
+    {
+        Dog = "Dog",
+        Cat = "Cat",
+        ZEBRA = 1234,
+    } as const, // the `as const` won't be required in TypeScript 5.0
+    "MyEnum" // friendly name is optional; used to generate helpful parser errors
+) 
+
+export type MyEnum = EnumOf<typeof MyEnum> // => "Dog" | "Cat" | 1234
+```
+
+## Methods
+
+### Standard enum-style accessors
+
+```ts
+const dog = MyEnum.Dog // => "Dog"
+const zebra = MyEnum.ZEBRA // => 1234
+```
+
+### Values
+
+Access array of all valid values, correctly typed (but without any ordering guarantees).
+
+```ts
+const myEnumValues = MyEnum.values // => ["Dog", "Cat", 1234]
+```
+
+### Parser
+
+Get a safe parser function for this enum automagically!
+
+```ts
+const parsed = MyEnum.parse("cat") // => a `Result<MyEnum, string>`, in this case `Result.ok("Cat")`
+```
+
+### Match
+
+See `match` in the {@link Variants} module docs for more details on matchers.
+
+```ts
+const speak: (e: MyEnum) => string = MyEnum.match({
+    Dog: "woof woof",
+    Cat: () => "meow",
+    ZEBRA: "is my skin white with black stripes or black with white stripes??",
+})
+
+speak(myEnum)
+```
+
+### MatchOrElse
+
+See `matchOrElse` in the {@link Variants} module docs for more details on partial matchers.
+
+```ts
+const speak: (e: MyEnum) => string = MyEnum.matchOrElse({
+    Dog: "woof woof",
+    Cat: () => "meow",
+    orELse: "I cannot speak.",
+})
+
+speak(myEnum)
+```
+
+@module Enums
+*/
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Result } from "./Result"
 import { String } from "./string"
 import { Option } from "./Option"
-import { pipe, flow } from "./composition"
+import { pipe, flow } from "./Composition"
 import { Array } from "./Array"
 import { Identity, NonNullish } from "./prelude"