+++ src/Result.ts
@@ -1,3 +1,8 @@
+import { Tagged, assertExhaustive, Refinement, NonNullish } from "./prelude"
+import { Option } from "./Option"
+import { flow, pipe } from "./Composition"
+import { EqualityComparer } from "./EqualityComparer"
+
 /**
  * The `Result` type represents the outcome of a completed operation
  * that either succeeded with some `Ok` value (also called a "success"
@@ -10,7 +15,7 @@
  * network request failed?)
  *
  * This API has been optimized for use with left-to-right function composition
- * using {@link composition/Pipe!pipe} and {@link composition/Flow!flow}.
+ * using {@link pipe} and {@link flow}.
  *
  * @example
  * ```
@@ -31,10 +36,6 @@
  * @module Result
  */
 /* eslint-disable @typescript-eslint/no-empty-interface */
-import { Tagged, assertExhaustive, Refinement, NonNullish } from "./prelude"
-import { Option } from "./Option"
-import { flow, pipe } from "./composition"
-import { EqualityComparer } from "./EqualityComparer"
 
 export interface Ok<A> extends Tagged<"Ok", { ok: A }> {}
 export interface Err<E> extends Tagged<"Err", { err: E }> {}