+++ src/composition/index.ts
@@ -1,2 +1,0 @@
-export { pipe } from "./Pipe"
-export { flow } from "./Flow"